Split specs into multiple files
