+++ spec/spec_helper.rb
@@ -1,6 +1,13 @@
 # frozen_string_literal: true
 
+require 'jekyll'
 require 'rspec-html-matchers'
+
+# Variables that should be shared across all specs in the suite.
+shared_context 'shared' do
+  @source_dir = File.join(__dir__, '..')
+  @dest_dir = File.join(@source_dir, '_site')
+end
 
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
@@ -104,6 +111,26 @@ RSpec.configure do |config|
   #   # test failures related to randomization by passing the same `--seed` value
   #   # as the one that triggered the failure.
   #   Kernel.srand config.seed
+
+  # Run 'jekyll build' before all specs in the suite
+  config.before(:suite) do
+    # TODO: Figure out how shared context can be used both in before:suite
+    # and in individual specs.
+    source_dir = File.join(__dir__, '..')
+    dest_dir = File.join(source_dir, '_site')
+    config = Jekyll.configuration(
+      {
+        'config' => File.join(source_dir, '_config.yml'),
+        'source' => source_dir,
+        'destination' => dest_dir
+      }
+    )
+    Jekyll::Commands::Build.process(config)
+  end
+
+  # Include the shared context
+  # config.include_context 'shared', :include_shared => true
+  # config.shared_context_metadata_behavior = :apply_to_host_groups
 end
 
 if ENV.fetch('GITHUB_ACTIONS', false) || ENV.fetch('COVER', false)
